Split AddMappings functions to 4 new files
