+++ cpp/src/gandiva/CMakeLists.txt
@@ -55,7 +55,10 @@ set(SRC_FILES
     annotator.cc
     bitmap_accumulator.cc
     cache.cc
-    cast_time.cc
+    gdv_time_function_stubs.cc
+    gdv_string_function_stubs.cc
+    gdv_hash_function_stubs.cc
+    gdv_arithmetic_function_stubs.cc
     configuration.cc
     context_helper.cc
     decimal_ir.cc
